+++ examples/bootstrap/webpack.config.js
@@ -64,7 +64,7 @@ module.exports = {
     ],
   },
 
-  // enable HMR with live reload
+  // enable live reload
   devServer: {
     static: path.resolve(__dirname, 'dist'),
     watchFiles: {