+++ examples/simple-site/webpack.config.js
@@ -53,7 +53,7 @@ module.exports = {
     ],
   },
 
-  // enable HMR with live reload
+  // enable live reload
   devServer: {
     static: path.resolve(__dirname, 'dist'),
     watchFiles: {