+++ test/cases/preprocessor-handlebars-helper-block-buildIn/webpack.config.js
@@ -1,0 +1,29 @@
+const path = require('path');
+const HtmlBundlerPlugin = require('../../../');
+
+const config = {
+  mode: 'production',
+  output: {
+    path: path.resolve(__dirname, './dist'),
+  },
+  plugins: [
+    new HtmlBundlerPlugin({
+      entry: {
+        index: {
+          import: './src/views/pages/home/index.hbs',
+          data: { title: 'Homepage' },
+        },
+        about: {
+          import: './src/views/pages/about/index.hbs',
+          data: { title: 'About' },
+        },
+      },
+      preprocessor: 'handlebars',
+      preprocessorOptions: {
+        partials: [path.join(__dirname, 'src/views/pages/'), path.join(__dirname, 'src/views/partials/')],
+      },
+    }),
+  ],
+};
+
+module.exports = config;