+++ test/jest.config.js
@@ -206,7 +206,7 @@ module.exports = {
   // testSequencer: '@jest/test-sequencer',
 
   // Default timeout of a test in milliseconds.
-  testTimeout: isLocalEnv ? 2000 : 5000,
+  testTimeout: isLocalEnv ? 2000 : 8000,
 
   // This option sets the URL for the jsdom environment. It is reflected in properties such as location.href
   // testURL: "http://localhost",